Refactor: "CardContent 파일 생성"
