+++ src/components/commons/CardContent/CardContent.module.scss
@@ -1,0 +1,55 @@
+.destination-card {
+  @include inline-flexbox($jc: center, $ai: start);
+  gap: 1rem;
+  border-radius: 1rem;
+  width: 45.2195rem;
+  height: 14.5374rem;
+  flex-shrink: 0;
+  box-shadow: 0rem 0.2rem 0.4rem rgba(0, 0, 0, 0.1);
+  padding: 1.24rem 0 1.27rem 1.08rem;
+
+  &__image-container {
+    width: 14.899rem;
+    height: 12.0194rem;
+    flex-shrink: 0;
+    overflow: hidden;
+    border-radius: 0.5rem;
+  }
+
+  &__image {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+  }
+
+  &__info {
+    @include column-flexbox($jc: start, $ai: start);
+    width: 100%;
+    gap: 4.06rem;
+  }
+
+  &__details {
+    @include column-flexbox($jc: start, $ai: start);
+    gap: 1rem;
+  }
+
+  &__name {
+    @include text-style(1.8, 700, #363636);
+  }
+
+  &__address {
+    @include text-style(1.3, 500, #363636);
+  }
+
+  &__ratings {
+    @include inline-flexbox($jc: center, $ai: center);
+    padding-left: 55%;
+    gap: 0.64rem;
+    flex-shrink: 0;
+  }
+
+  &__rating-item {
+    @include inline-flexbox($jc: center, $ai: center);
+    gap: 0.1rem;
+  }
+}
